requirements.txt
GitPython
Pillow
liturgical-calendar @ git+https://github.com/ludwigw/liturgical-calendar.git 